+++ src/Core/Spec/Spec12.php
@@ -116,4 +116,9 @@ final class Spec12 implements SpecInterface
     {
         return true;
     }
+
+    public function supportsToolExternalReferences(): bool
+    {
+        return false;
+    }
 }